requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain-google-genai
pillow
sentence-transformers
faiss-cpu
numpy
pysqlite3-binary
langchain-huggingface
chromadb
cassandra-driver
python-dotenv
langchain-astradb